[권한설정 마무리],[업무일지 작업중] 권한설정 방법은 내일 모니위키에 올리겠습니다

+++ groupware/src/main/java/com/sp/authority/Authority.java
@@ -12,13 +12,12 @@ public class Authority {
 	private String comments;
 	private MultipartFile upload;
 	private String saveFilename;
-	private long grants;
+	private int grants;
 	
-
-	public long getGrants() {
+	public int getGrants() {
 		return grants;
 	}
-	public void setGrants(long grants) {
+	public void setGrants(int grants) {
 		this.grants = grants;
 	}
 	public MultipartFile getUpload() {
